Crate actix_redis

Redis integration for Actix and session store for Actix Web.

Structs

Command
Command for send data to Redis

RedisActor
Redis comminucation actor

RedisSession
Use redis as session storage.

Enums

Error
General purpose actix redis error

RespError

RespValue
A single RESP value, this owns the data that is read/to-be written to Redis.

SameSite
The SameSite cookie attribute.