Crate actix_redis

Redis integration for actix.

Macros

resp_array

Macro to create a RESP array, useful for preparing commands to send. Elements can be any type, or a mixture of types, that satisfy Into<RespValue>.

Structs

Command

Command for send data to Redis

RedisActor

Redis communication actor

Enums

Error

General purpose actix-redis error.

RespError

RespValue

A single RESP value, this owns the data that is read/to-be written to Redis.