Enum actix_redis::RespValue

pub enum RespValue {
    Nil,
    Array(Vec<RespValue>),
    BulkString(Vec<u8>),
    Error(String),
    Integer(i64),
    SimpleString(String),
}

A single RESP value, this owns the data that is read/to-be written to Redis.

It is cloneable to allow multiple copies to be delivered in certain circumstances, e.g. multiple subscribers to the same topic.

Variants
NilArray(Vec<RespValue>)

Zero, one or more other RespValues.

BulkString(Vec<u8>)

A bulk string. In Redis terminology a string is a byte-array, so this is stored as a vector of u8s to allow clients to interpret the bytes as appropriate.

Error(String)

An error from the Redis server

Integer(i64)

Redis documentation defines an integer as being a signed 64-bit integer: https://redis.io/topics/protocol#resp-integers

SimpleString(String)